refactor: dedupe week-nav helpers across notification tabs
getMonday/formatWeekLabel were copy-pasted into AuditLogTab, RecentActivityTab,
and NewCasesTab. Moved to lib/formatters.ts as the single source; formatWeekLabel
now takes a spanDays param (4 for Mon-Fri, 6 for New Cases' full Mon-Sun week).
