+++ src/components/notifications/NewCasesTab.tsx
@@ -4,6 +4,7 @@ import { useState, useEffect, useRef, Fragment } from "react";
 import Link from "next/link";
 import { ChevronLeft, ChevronRight, RefreshCw, UserPlus, AlertCircle, ExternalLink } from "lucide-react";
 import { themeClasses } from "@/lib/theme-classes";
+import { getMonday, formatWeekLabel as formatWeekLabelBase } from "@/lib/formatters";
 
 interface DayCount {
   date: string;
@@ -23,20 +24,10 @@ interface NewCasesTabProps {
   t: ReturnType<typeof themeClasses>;
 }
 
-const getMonday = (offset = 0): string => {
-  const d = new Date();
-  const day = d.getDay();
-  d.setDate(d.getDate() - day + (day === 0 ? -6 : 1) + offset * 7);
-  return d.toISOString().split("T")[0];
-};
-
-const formatWeekLabel = (monday: string): string => {
-  const start = new Date(monday + "T00:00:00");
-  const end = new Date(monday + "T00:00:00");
-  end.setDate(end.getDate() + 6);
-  const opts: Intl.DateTimeFormatOptions = { month: "short", day: "numeric" };
-  return `${start.toLocaleDateString("en-US", opts)} – ${end.toLocaleDateString("en-US", { ...opts, year: "numeric" })}`;
-};
+// New cases land every day, not just weekdays, so this tab spans the full
+// Mon-Sun week (6 days past Monday) rather than the Mon-Fri default used by
+// Audit Log / Recent Activity.
+const formatWeekLabel = (monday: string): string => formatWeekLabelBase(monday, 6);
 
 const fmtDate = (iso: string): string =>
   new Date(`${iso}T00:00:00`).toLocaleDateString("en-US", {
